refactor: Use `instanceof` approach instead
